requirements.txt
Django==5.0.0
django-tenants==3.6.1
psycopg[binary]==3.1.13
django-redis==5.4.0
celery==5.3.4
django-cors-headers==4.3.1
python-decouple==3.8
whitenoise==6.6.0
redis==5.0.1
djangorestframework==3.14.0
django-guardian==2.4.0
uvicorn==0.24.0
# Pipeline system dependencies
pydantic==2.5.0
django-filter==23.5
jsonschema==4.20.0
python-dateutil==2.8.2
pillow==10.1.0
phonenumbers==8.13.25
openpyxl==3.1.2
# AI integration
openai==1.6.0
cryptography>=45.0.0
# Phase 05 - API Layer dependencies
strawberry-graphql==0.278.0
strawberry-graphql-django==0.65.1
django-filter==23.5
drf-spectacular==0.27.0
channels-redis==4.2.0
django-ratelimit==4.1.0
drf-nested-routers==0.93.4
channels==4.0.0

# Phase 07 - Workflow Automation dependencies
croniter==1.4.1
aiohttp==3.9.1
django-celery-beat==2.5.0
python-crontab==3.0.0